renamed InputHandler to InputManager. Added some AI entity logging. Its very spammy, but needed for now.
