+++ src/core/GameEngine.cpp
@@ -12,7 +12,7 @@
 #include "managers/FontManager.hpp"
 #include "states/GamePlayState.hpp"
 #include "managers/GameStateManager.hpp"
-#include "core/InputHandler.hpp"
+#include "managers/InputManager.hpp"
 #include "states/LogoState.hpp"
 #include "states/MainMenuState.hpp"
 #include "SDL3/SDL_render.h"
